fix: split treasury details widgets under 300-line gate
Quality failed repo.game_widgets_file_size on the 411-line support file; extract format helpers and panel so each shell module stays within the ceiling.
